Moving point light ray marching scene

+++ app/src/main/java/com/inasweaterpoorlyknit/learnopengl_androidport/utils/SystemUtils.kt
@@ -1,10 +1,10 @@
 package com.inasweaterpoorlyknit.learnopengl_androidport.utils
 
-fun systemTimeInSeconds() : Float {
-    return System.nanoTime().toFloat() / 1000000000.0f
+fun systemTimeInSeconds() : Double {
+    return System.nanoTime().toDouble() / 1000000000
 }
 
-fun systemTimeInDeciseconds() : Float {
+fun systemTimeInDeciseconds() : Double {
     // note: time measured in deciseconds (10^-1 seconds)
-    return System.nanoTime().toFloat() / 100000000.0f
+    return System.nanoTime().toDouble() / 100000000
 }
